add optional manager param exit_on_load
A system with hundreds of nodelets can waste a lot of resources
managing bonds and just waking up to check if shutdown. Add a manager
parameter 'exit_on_load' to have all nodelet loaders exit after
loading the nodelet successfully to the manager. This is mainly
useful when there is no need to unload the nodelet apart from exiting
the manager. Also, if 'exit_on_load' is set, if any nodelet fails to
load, the manager exits. This way the manager can be 'required' in a
roslaunch file and failure can be detected even though the nodelet
load failed, as the load can not be required since it will exit.